Refactored InvestmentController with new endpoints. Added skeleton unit tests for InvestmentController.

+++ src/main/java/com/citi/training/PortfolioManager/service/InvestmentService.java
@@ -21,5 +21,5 @@ public interface InvestmentService {
 
     Double getInvestmentCurrentPrice(Investment investment);
 
-    Integer getTotalInvestmentValue();
+    Double getTotalInvestmentValue();
 }
